requirements.txt
python-telegram-bot==20.7
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
openpyxl